--- .env.example
@@ -1,1 +1,0 @@
-LIB_NAME=itg
