+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "itg",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "description": "Issue Tracker for Github",
   "main": "bin.js",
   "bin": "./bin.js",
@@ -24,7 +24,6 @@
   "license": "ISC",
   "homepage": "https://github.com/Pawelek99/itg#readme",
   "dependencies": {
-    "dotenv": "^8.2.0",
     "git-root-dir": "^1.0.2",
     "node-fetch": "^2.6.0",
     "read": "^1.0.7",
@@ -39,13 +38,16 @@
     "src/issue.js",
     "src/pr.js",
     "src/utils.js",
-    "src/api.js",
     "src/cache.js",
+    "src/api/api.js",
+    "src/api/grapql.js",
+    "src/api/utils.js",
     "README.md"
   ],
   "prettier": {
     "trailingComma": "all",
     "singleQuote": true,
-    "tabWidth": 2
+    "tabWidth": 2,
+    "useTabs": true
   }
 }
