+++ src/cache.js
@@ -6,108 +6,108 @@ let loaded = false;
 let path = undefined;
 
 const computePath = async () => {
-  if (path) {
-    return;
-  }
+	if (path) {
+		return;
+	}
 
-  const git = await gitRootDir();
+	const git = await gitRootDir();
 
-  if (!git) {
-    sh.echo('You can only use this library in the git directory');
-    sh.exit(1);
-  }
+	if (!git) {
+		sh.echo('You can only use this library in the git directory');
+		sh.exit(1);
+	}
 
-  path = `${git}/.git/.itg.cache`;
+	path = `${git}/.git/.itg.cache`;
 };
 
 const clearFile = async () => {
-  await computePath();
+	await computePath();
 
-  fs.writeFileSync(path, '');
+	fs.writeFileSync(path, '');
 };
 
 const loadFile = async () => {
-  await computePath();
+	await computePath();
 
-  // Avoid loading file if it's already loaded
-  if (loaded || !fs.existsSync(path)) {
-    return;
-  }
+	// Avoid loading file if it's already loaded
+	if (loaded || !fs.existsSync(path)) {
+		return;
+	}
 
-  loaded = true;
+	loaded = true;
 
-  // Load file contents and treat each line as a 'key:value' pair
-  const content = fs.readFileSync(path).toLocaleString();
-  content.split('\n').forEach((line) => {
-    if (!line) {
-      return;
-    }
+	// Load file contents and treat each line as a 'key:value' pair
+	const content = fs.readFileSync(path).toLocaleString();
+	content.split('\n').forEach((line) => {
+		if (!line) {
+			return;
+		}
 
-    const pair = line.split(':');
-    if (pair[0] === '__UNTIL__') {
-      if (Date.now() - pair[1] >= 0) {
-        cache.clear();
-        clearFile();
-        return;
-      }
-    }
-    cache.set(Buffer.from(pair[0], 'base64').toString('utf-8'), pair[1]);
-  });
+		const pair = line.split(':');
+		if (pair[0] === '__UNTIL__') {
+			if (Date.now() - pair[1] >= 0) {
+				cache.clear();
+				clearFile();
+				return;
+			}
+		}
+		cache.set(Buffer.from(pair[0], 'base64').toString('utf-8'), pair[1]);
+	});
 };
 
 const saveFile = async () => {
-  await computePath();
+	await computePath();
 
-  const data = [];
-  let until;
+	const data = [];
+	let until;
 
-  // Save cache entries to an array as 'key:value' pair and then join them with a newline character
-  cache.forEach((value, key) => {
-    if (key === '__UNTIL__') {
-      until = value;
-    } else {
-      data.push(`${Buffer.from(key).toString('base64')}:${value}`);
-    }
-  });
+	// Save cache entries to an array as 'key:value' pair and then join them with a newline character
+	cache.forEach((value, key) => {
+		if (key === '__UNTIL__') {
+			until = value;
+		} else {
+			data.push(`${Buffer.from(key).toString('base64')}:${value}`);
+		}
+	});
 
-  if (!until) {
-    const threeDays = 1000 * 60 * 60 * 24 * 3;
-    until = Date.now() + threeDays;
-  }
+	if (!until) {
+		const threeDays = 1000 * 60 * 60 * 24 * 3;
+		until = Date.now() + threeDays;
+	}
 
-  data.push(`__UNTIL__:${until}`);
+	data.push(`__UNTIL__:${until}`);
 
-  fs.writeFileSync(path, data.join('\n'));
+	fs.writeFileSync(path, data.join('\n'));
 };
 
 const get = async (key) => {
-  await loadFile();
+	await loadFile();
 
-  const value = cache.get(key);
+	const value = cache.get(key);
 
-  if (!value) {
-    return;
-  }
+	if (!value) {
+		return;
+	}
 
-  return JSON.parse(Buffer.from(value, 'base64').toString('utf-8'));
+	return JSON.parse(Buffer.from(value, 'base64').toString('utf-8'));
 };
 
 const set = async (key, value) => {
-  cache.set(key, Buffer.from(JSON.stringify(value)).toString('base64'));
+	cache.set(key, Buffer.from(JSON.stringify(value)).toString('base64'));
 
-  await saveFile();
+	await saveFile();
 
-  return value;
+	return value;
 };
 
 const has = async (key) => {
-  await loadFile();
+	await loadFile();
 
-  return cache.has(key);
+	return cache.has(key);
 };
 
 module.exports = {
-  get,
-  set,
-  has,
+	get,
+	set,
+	has,
 };
